Starting implementation of Strategy Pattern on Patterns
