+++ bin/.gitignore
@@ -6,3 +6,4 @@
 /tests/
 /textboxes/
 /themes/
+/patterns/
