+++ output.txt
@@ -1,52 +1,54 @@
-5-pointed_star 105 0 105 105 0 -1 
-5-pointed_star 315 0 105 105 0 -1 
-5-pointed_star 525 0 105 105 0 -1 
-5-pointed_star 735 0 105 105 0 -1 
-5-pointed_star 945 0 105 105 0 -1 
-5-pointed_star 0 105 105 105 0 -1 
-5-pointed_star 210 105 105 105 0 -1 
-5-pointed_star 420 105 105 105 0 -1 
-5-pointed_star 630 105 105 105 0 -1 
-5-pointed_star 840 105 105 105 0 -1 
-5-pointed_star 105 210 105 105 0 -1 
-5-pointed_star 315 210 105 105 0 -1 
-5-pointed_star 525 210 105 105 0 -1 
-5-pointed_star 735 210 105 105 0 -1 
-5-pointed_star 945 210 105 105 0 -1 
-5-pointed_star 0 315 105 105 0 -1 
-5-pointed_star 210 315 105 105 0 -1 
-5-pointed_star 420 315 105 105 0 -1 
-5-pointed_star 630 315 105 105 0 -1 
-5-pointed_star 840 315 105 105 0 -1 
-5-pointed_star 105 420 105 105 0 -1 
-5-pointed_star 315 420 105 105 0 -1 
-5-pointed_star 525 420 105 105 0 -1 
-5-pointed_star 735 420 105 105 0 -1 
-5-pointed_star 945 420 105 105 0 -1 
-5-pointed_star -1 -1 105 105 0 -1 
-6-pointed_star 0 0 105 105 0 -1 
-6-pointed_star 210 0 105 105 0 -1 
-6-pointed_star 420 0 105 105 0 -1 
-6-pointed_star 630 0 105 105 0 -1 
-6-pointed_star 840 0 105 105 0 -1 
-6-pointed_star 105 105 105 105 0 -1 
-6-pointed_star 315 105 105 105 0 -1 
-6-pointed_star 525 105 105 105 0 -1 
-6-pointed_star 735 105 105 105 0 -1 
-6-pointed_star 945 105 105 105 0 -1 
-6-pointed_star 0 210 105 105 0 -1 
-6-pointed_star 210 210 105 105 0 -1 
-6-pointed_star 420 210 105 105 0 -1 
-6-pointed_star 630 210 105 105 0 -1 
-6-pointed_star 840 210 105 105 0 -1 
-6-pointed_star 105 315 105 105 0 -1 
-6-pointed_star 315 315 105 105 0 -1 
-6-pointed_star 525 315 105 105 0 -1 
-6-pointed_star 735 315 105 105 0 -1 
-6-pointed_star 945 315 105 105 0 -1 
-6-pointed_star 0 420 105 105 0 -1 
-6-pointed_star 210 420 105 105 0 -1 
-6-pointed_star 420 420 105 105 0 -1 
-6-pointed_star 630 420 105 105 0 -1 
-6-pointed_star 840 420 105 105 0 -1 
-6-pointed_star -1 -1 105 105 0 -1 
+circle 105 0 105 105 0 -3670016 
+circle 315 0 105 105 0 -3670016 
+circle 525 0 105 105 0 -3670016 
+circle 735 0 105 105 0 -3670016 
+circle 945 0 105 105 0 -3670016 
+circle 0 105 105 105 0 -3670016 
+circle 210 105 105 105 0 -3670016 
+circle 420 105 105 105 0 -3670016 
+circle 630 105 105 105 0 -3670016 
+circle 840 105 105 105 0 -3670016 
+circle 105 210 105 105 0 -3670016 
+circle 315 210 105 105 0 -3670016 
+circle 525 210 105 105 0 -3670016 
+circle 735 210 105 105 0 -3670016 
+circle 945 210 105 105 0 -3670016 
+circle 0 315 105 105 0 -3670016 
+circle 210 315 105 105 0 -3670016 
+circle 420 315 105 105 0 -3670016 
+circle 630 315 105 105 0 -3670016 
+circle 840 315 105 105 0 -3670016 
+circle 105 420 105 105 0 -3670016 
+circle 315 420 105 105 0 -3670016 
+circle 525 420 105 105 0 -3670016 
+circle 735 420 105 105 0 -3670016 
+circle 945 420 105 105 0 -3670016 
+circle -1 -1 105 105 0 -3670016 
+circle 105 0 105 105 0 -3670016 
+square 0 0 105 105 0 -3670016 
+square 210 0 105 105 0 -3670016 
+square 420 0 105 105 0 -3670016 
+square 630 0 105 105 0 -3670016 
+square 840 0 105 105 0 -3670016 
+square 105 105 105 105 0 -3670016 
+square 315 105 105 105 0 -3670016 
+square 525 105 105 105 0 -3670016 
+square 735 105 105 105 0 -3670016 
+square 945 105 105 105 0 -3670016 
+square 0 210 105 105 0 -3670016 
+square 210 210 105 105 0 -3670016 
+square 420 210 105 105 0 -3670016 
+square 630 210 105 105 0 -3670016 
+square 840 210 105 105 0 -3670016 
+square 105 315 105 105 0 -3670016 
+square 315 315 105 105 0 -3670016 
+square 525 315 105 105 0 -3670016 
+square 735 315 105 105 0 -3670016 
+square 945 315 105 105 0 -3670016 
+square 0 420 105 105 0 -3670016 
+square 210 420 105 105 0 -3670016 
+square 420 420 105 105 0 -3670016 
+square 630 420 105 105 0 -3670016 
+square 840 420 105 105 0 -3670016 
+square -1 -1 105 105 0 -3670016 
+square 0 0 105 105 0 -3670016
