+++ src/patterns/Pattern.java
@@ -1,0 +1,9 @@
+package patterns;
+
+public abstract class Pattern {
+	protected static int width;
+	protected static int height;
+	
+	
+
+}
